alerts added. Need customized

+++ client/src/components/FlashCards/index.js
@@ -39,6 +39,8 @@ const FlashCards = (props) => {
         },
       });
       closeModal();
+        // alert user that flash card was updated and re load the page so that it shows
+      window.alert('Flash Card Updated!').location.reload();
     } catch (err) {
       console.error(err);
     }
@@ -46,18 +48,23 @@ const FlashCards = (props) => {
 
   // delete flash card function
   const handleDeleteFlashCard = async () => {
-    try {
-      await deleteFlashCard({
-        variables: {
-          id: currentFlashCard._id,
-        },
-      });
-      closeModal();
-    } catch (err) {
-      console.error(err);
+    const confirmDelete = window.confirm('Are you sure you want to delete this flash card?');
+    if (confirmDelete) {
+      try {
+        await deleteFlashCard({
+          variables: {
+            id: currentFlashCard._id,
+          },
+        });
+        // close the modal
+        closeModal()
+        // alert user that flash card was deleted and re load the page so that it shows
+        window.alert('Flash Card Deleted!').location.reload();
+      } catch (err) {
+        console.error(err);
+      }
     }
   };
-
   const { loading, data } = useQuery(VIEW_FLASHCARDS);
   // viewFlashCards is from the query / graphql
   const flashcards = props.flashCards || [];
